adding driver choice for intermediate files
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rspatial@1011 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ spgrass6/R/vect_link.R
@@ -1,14 +1,15 @@
-# Interpreted GRASS 6 interface functions
-# Copyright (c) 2005-9 Roger S. Bivand
+# Interpreted GRASS 6+ interface functions
+# Copyright (c) 2005-2010 Roger S. Bivand
 #
 readVECT6 <- function(vname, layer, type=NULL, plugin=NULL,
         remove.duplicates=TRUE, 
 	ignore.stderr = FALSE, with_prj=TRUE, with_c=FALSE, mapset=NULL, 
-	pointDropZ=FALSE) {
+	pointDropZ=FALSE, driver="ESRI Shapefile") {
 
     G7 <- execGRASS("g.version", intern=TRUE) > "GRASS 7"
     if (missing(layer)) layer <- 1L
     if (G7) layer <- as.character(layer)
+    if (driver == "GRASS") plugin <- TRUE
 
     if (is.null(plugin)) {
         ogrD <- ogrDrivers()$name
@@ -38,6 +39,20 @@ readVECT6 <- function(vname, layer, type=NULL, plugin=NULL,
                 verbose=!ignore.stderr)
         }
     } else {
+        ogrD <- ogrDrivers()$name
+	if (!(driver %in% ogrD))
+            stop(paste("Requested driver", driver, "not available in rgdal"))
+        ogrDGRASS <- execGRASS("v.in.ogr", flags="f", intern=TRUE,
+            ignore.stderr=ignore.stderr)
+        ogrDGRASSs <- strsplit(ogrDGRASS, ": ")
+        if (!(driver %in% sapply(ogrDGRASSs, "[", 2)))
+            stop(paste("Requested driver", driver, "not available in GRASS"))
+        vDrivers <- c("ESRI_Shapefile", "MapInfo_File", "TIGER", "S57",
+            "DGN", "Memory", "BNA", "CSV", "GML", "GPX", "KML", "GeoJSON",
+            "Interlis_1", "Interlis_2", "GMT", "SQLite", "ODBC",
+            "PostgreSQL", "MySQL")
+        if (!(gsub(" ", "_", driver) %in% vDrivers))
+            stop(paste("Requested driver", driver, "not valid for v.out.ogr"))
 	vinfo <- vInfo(vname)
 	types <- names(vinfo)[which(vinfo > 0)]
 	if (is.null(type)) {
@@ -64,7 +79,7 @@ readVECT6 <- function(vname, layer, type=NULL, plugin=NULL,
         if (with_c) flags <- c(flags, "c")
         execGRASS("v.out.ogr", flags=flags, parameters=list(input=vname,
             type=type, layer=layer, dsn=gtmpfl1, olayer=shname,
-            format="ESRI_Shapefile"), ignore.stderr=ignore.stderr)
+            format=gsub(" ", "_", driver)), ignore.stderr=ignore.stderr)
 
         if (sss[1] >= "0.6" && as.integer(sss[2]) > 7) {
 	    res <- readOGR(dsn=rtmpfl1, layer=shname, verbose=!ignore.stderr, 
@@ -72,9 +87,16 @@ readVECT6 <- function(vname, layer, type=NULL, plugin=NULL,
         } else {
 	    res <- readOGR(dsn=rtmpfl1, layer=shname, verbose=!ignore.stderr)           }
 
-	if (.Platform$OS.type != "windows") 
-            unlink(paste(rtmpfl1, list.files(rtmpfl1, pattern=shname), 
-		sep=.Platform$file.sep))
+        if (sss[1] >= "0.6") {
+            if (sss[1] < "0.7" && as.integer(sss[2]) < 32) {
+	        if (.Platform$OS.type != "windows") 
+                    unlink(paste(rtmpfl1, list.files(rtmpfl1, pattern=shname), 
+		        sep=.Platform$file.sep))
+            } else {
+                rgdal:::ogrDeleteLayer(dsn=rtmpfl1, layer=shname)
+            }
+        }
+        
 	if (remove.duplicates && type != "point") {
 		dups <- duplicated(slot(res, "data"))
 		if (any(dups)) {
@@ -187,8 +209,17 @@ readVECT6 <- function(vname, layer, type=NULL, plugin=NULL,
 }
 
 writeVECT6 <- function(SDF, vname, #factor2char = TRUE, 
-	v.in.ogr_flags=NULL, ignore.stderr = FALSE) {
+	v.in.ogr_flags=NULL, ignore.stderr = FALSE, driver="ESRI Shapefile") {
 
+        ogrD <- ogrDrivers()$name
+	if (!(driver %in% ogrD))
+            stop(paste("Requested driver", driver, "not available in rgdal"))
+        ogrDGRASS <- execGRASS("v.in.ogr", flags="f", intern=TRUE,
+            ignore.stderr=ignore.stderr)
+        ogrDGRASSs <- strsplit(ogrDGRASS, ": ")
+        if (!(driver %in% sapply(ogrDGRASSs, "[", 2)))
+            stop(paste("Requested driver", driver, "not available in GRASS"))
+        sss <- strsplit(packageDescription("rgdal")$Version, "-")[[1]]
 	type <- NULL
 	if (class(SDF) == "SpatialPointsDataFrame") type <- "point"
 	if (class(SDF) == "SpatialLinesDataFrame") type <- "line"
@@ -207,16 +238,22 @@ writeVECT6 <- function(SDF, vname, #factor2char = TRUE,
 		nchar(vname)))
 
 
-	writeOGR(SDF, dsn=rtmpfl1, layer=shname, driver="ESRI Shapefile")
+	writeOGR(SDF, dsn=rtmpfl1, layer=shname, driver=driver)
 
 
 	execGRASS("v.in.ogr", flags=v.in.ogr_flags,
 	    parameters=list(dsn=gtmpfl1, output=vname, layer=shname),
 	    ignore.stderr=ignore.stderr)
 
-	if (.Platform$OS.type != "windows") 
-            unlink(paste(rtmpfl1, list.files(rtmpfl1, pattern=shname), 
-		sep=.Platform$file.sep))
+        if (sss[1] >= "0.6") {
+            if (sss[1] < "0.7" && as.integer(sss[2]) < 32) {
+	        if (.Platform$OS.type != "windows") 
+                    unlink(paste(rtmpfl1, list.files(rtmpfl1, pattern=shname), 
+		        sep=.Platform$file.sep))
+            } else {
+                rgdal:::ogrDeleteLayer(dsn=rtmpfl1, layer=shname)
+            }
+        }
 
 }
